+++ brandon-graf-frontend/src/appHeight.jsx
@@ -1,0 +1,10 @@
+const appHeight = () =>  {
+    const doc = document.documentElement
+    doc.style.setProperty('--app-height', `${window.innerHeight}px`)
+    console.log(window.innerHeight)
+}
+
+window.addEventListener('resize', appHeight)
+appHeight()
+
+export default appHeight
